pyproject.toml
[tool.poetry]
name = "UDT_cINN"
version = "0.0.1"
description = "This is the project repo for the unsupervised domain transfer project submitted to miccai23"
authors = ["Kris Dreher <k.dreher@dkfz-heidelberg.de>", "Leonardo Ayala <l.menjivar@dkfz-heidelberg.de>"]
license = "MIT"
readme = "README.md"

packages = [
    { include = "src" },
]

# Requirements
[tool.poetry.dependencies]
python = ">=3.8"
FrEIA = { git = "https://github.com/VLL-HD/FrEIA.git" }
click = "*"
Sphinx = "*"
coverage = "*"
flake8 = "*"
python-dotenv = ">=0.5.1"
pandas = "*"
numpy  = "<1.24"
scikit-learn = "*"
scipy = "*"
pytorch-lightning = "1.9.0"
omegaconf = "*"
matplotlib_scalebar = "*"
openpyxl = "*"
torch-fidelity = "*"
tensorboard = "*"
seaborn = "*"
dask = "^2023.1.0"
plotly = "^5.13.1"
kaleido = "*"

[tool.poetry.group.docs.dependencies]
sphinx-rtd-theme = "^1.0.0"
Sphinx = "^5.1.1"
myst-parser = "^0.18.0"

[build-system]
requires = [
    "poetry >= 0.12"
]
build-backend = "poetry.masonry.api"